# specs/001-web-automation-checker/contracts/api.yaml
openapi: 3.0.3
info:
  title: Web Automation Checker API
  description: |
    REST API for the Web Automation Checker tool. Allows users to submit URLs for automated testing,
    retrieve test reports, and view historical test data.

    **Features**:
    - P1: Basic URL testing with UI and performance checks
    - P2: Detailed diagnostic information for failures
    - P3: Test history and trend analysis
  version: 1.0.0
  contact:
    name: Anita Project Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://checker.example.com/api/v1
    description: Production server
...
paths:
  /tests:
    post:
      tags:
        - Tests
      summary: Create a new test request
      description: |
        Submit a URL for automated testing. The system will perform UI functionality checks
        and performance measurements, then return a test report.

        **User Story**: P1 - Basic URL Testing
      operationId: createTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequest'
            examples:
              basic:
                summary: Basic URL test
                value:
                  url: "https://example.com"
              with_config:
                summary: Test with custom configuration
                value:
                  url: "https://example.com"
                  config:
                    timeout: 45
                    waitTime: 7
      responses:
        '201':
          description: Test request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequest'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                url: "https://example.com"
                requestedAt: "2025-11-27T10:30:00Z"
                status: "pending"
                config:
                  timeout: 30
                  waitTime: 5
        '400':
          description: Invalid request (bad URL format, invalid config)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Validation Error"
                message: "Invalid URL format. Must be http:// or https://"
                details:
                  - field: "url"
                    issue: "Must match pattern ^https?://"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Rate Limit Exceeded"
                message: "Too many requests. Max 10 per minute."
                retryAfter: 45
  /tests/{testId}:
    get:
      tags:
        - Tests
      summary: Get test request status
      description: |
        Retrieve the current status of a test request. Poll this endpoint to monitor
        test execution progress.
      operationId: getTestStatus
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Test request ID
      responses:
        '200':
          description: Test request found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequest'
        '404':
          description: Test request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reports/{reportId}:
    get:
      tags:
        - Reports
      summary: Get test report by ID
      description: |
        Retrieve complete test results including UI test results and performance metrics.

        **User Stories**: P1 (Basic results) + P2 (Detailed diagnostics)
      operationId: getReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Test report ID
      responses:
        '200':
          description: Test report found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestReport'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reports:
    get:
      tags:
        - Reports
      summary: List recent test reports
      description: |
        Retrieve a paginated list of recent test reports.
      operationId: listReports
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of reports to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of reports to skip
        - name: url
          in: query
          schema:
            type: string
            format: uri
          description: Filter by tested URL
      responses:
        '200':
          description: Reports retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestReportSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  schemas:
    CreateTestRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          pattern: '^https?://'
          maxLength: 2048
          description: URL to test
          example: "https://example.com"
        config:
          type: object
          properties:
            timeout:
              type: integer
              minimum: 5
              maximum: 60
              default: 30
              description: Test timeout in seconds
            waitTime:
              type: integer
              minimum: 0
              maximum: 10
              default: 5
              description: Wait time for JavaScript execution (seconds)
    TestRequest:
      type: object
      required:
        - id
        - url
        - requestedAt
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique test request identifier
        url:
          type: string
          format: uri
          description: The URL being tested
        requestedAt:
          type: string
          format: date-time
          description: When the test was requested
        status:
          type: string
          enum: [pending, running, completed, failed]
          description: Current status of the test
        config:
          type: object
          properties:
            timeout:
              type: integer
            waitTime:
              type: integer
    TestReport:
      type: object
      required:
        - id
        - testRequestId
        - url
        - overallScore
        - totalChecks
        - passedChecks
        - failedChecks
        - warningChecks
        - testDuration
        - completedAt
        - uiTestResults
        - performanceResults
      properties:
        id:
          type: string
          format: uuid
        testRequestId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall health score (0-100)
        totalChecks:
          type: integer
          minimum: 0
        passedChecks:
          type: integer
          minimum: 0
        failedChecks:
          type: integer
          minimum: 0
        warningChecks:
          type: integer
          minimum: 0
        testDuration:
          type: integer
          minimum: 0
          description: Test execution time in milliseconds
        completedAt:
          type: string
          format: date-time
        uiTestResults:
          type: array
          items:
            $ref: '#/components/schemas/UITestResult'
        performanceResults:
          type: array
          items:
            $ref: '#/components/schemas/PerformanceResult'
    TestReportSummary:
      type: object
      description: Lightweight version of TestReport for list views
      required:
        - id
        - url
        - overallScore
        - completedAt
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
        totalChecks:
          type: integer
        passedChecks:
          type: integer
        failedChecks:
          type: integer
        completedAt:
          type: string
          format: date-time
    UITestResult:
      type: object
      required:
        - id
        - testReportId
        - testType
        - status
      properties:
        id:
          type: string
          format: uuid
        testReportId:
          type: string
          format: uuid
        testType:
          type: string
          enum: [link, form, button, image]
          description: Type of UI element tested
        elementId:
          type: string
          maxLength: 255
          description: CSS selector or element description
        status:
          type: string
          enum: [pass, fail, warning]
        errorMessage:
          type: string
          maxLength: 1000
          description: Error description if test failed
        screenshotUrl:
          type: string
          format: uri
          description: Screenshot of failed element (P2 feature)
        recommendation:
          type: string
          maxLength: 500
          description: Suggested fix (P2 feature)
        diagnostics:
          type: object
          description: Additional debugging information
          additionalProperties: true
    PerformanceResult:
      type: object
      required:
        - id
        - testReportId
        - metricName
        - measuredValue
        - unit
        - threshold
        - status
      properties:
        id:
          type: string
          format: uuid
        testReportId:
          type: string
          format: uuid
        metricName:
          type: string
          enum: [loadTime, resourceSize, responseTime, renderTime]
        measuredValue:
          type: number
          format: double
          minimum: 0
        unit:
          type: string
          enum: [ms, bytes, score]
        threshold:
          type: number
          format: double
          minimum: 0
        status:
          type: string
          enum: [pass, fail, warning]
        recommendation:
          type: string
          maxLength: 500
          description: Optimization suggestion (P2 feature)
        details:
          type: object
          description: Detailed breakdown (e.g., largest resources)
          additionalProperties: true
...
    Pagination:
      type: object
      required:
        - total
        - limit
        - offset
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Current offset
        hasMore:
          type: boolean
          description: Whether more items are available
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
          description: Validation error details
        retryAfter:
          type: integer
          description: Seconds to wait before retrying (for rate limits)
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future enhancement)
security: []  # No authentication required for MVP